@@ -17,26 +17,27 @@ environment:
   sdk: ">=2.1.0 <3.0.0"
 
 dependencies:
-  flare_splash_screen: ^3.0.1
-  shared_preferences: ^0.5.6+3
-  cloud_firestore: ^0.13.4
-  file_picker: ^1.5.1
+  flare_splash_screen: ^4.0.0
+  shared_preferences: ^2.0.15
+  cloud_firestore: ^3.1.17
+  file_picker: ^4.6.1
   open_file: ^3.0.1
-  flare_flutter: ^2.0.3
-  firebase_messaging: 6.0.13
-  http: ^0.12.0+4
-  flutter_launcher_icons: ^0.8.1
-  page_transition: "^1.1.5"
-  cached_network_image: ^2.2.0
-  photo_view: ^0.9.2
-  google_fonts: ^1.1.0 
-  uuid: ^2.0.4
-  url_launcher: ^5.7.10
-  path_provider: ^1.6.11
-  fluttertoast: ^7.1.1
+  flare_flutter: ^3.0.2
+  firebase_messaging: ^11.4.1
+  http: ^0.13.0
+  flutter_launcher_icons: ^0.9.3
+  page_transition: ^2.0.5
+  cached_network_image: ^3.2.1
+  photo_view: ^0.14.0
+  google_fonts: ^3.0.1
+  uuid: ^3.0.6
+  url_launcher: ^6.1.3
+  path_provider: ^2.0.11
+  fluttertoast: ^8.0.9
   progress_dialog: ^1.2.4
-  firebase_storage: 3.0.4
-  youtube_parser: ^1.1.0
+  firebase_storage: ^10.2.16
+  youtube_parser: ^2.0.0
+  firebase_storage_web: ^3.2.15
 
   flutter:
     sdk: flutter
@@ -44,7 +45,7 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.5
 
 dev_dependencies:
   flutter_test:
