@@ -6,7 +6,7 @@ import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:progress_dialog/progress_dialog.dart';
-
+import 'Utils.dart';
 bool isActive = true;
 String userRole = "3";
 bool loading = false;
@@ -48,10 +48,10 @@ class _AddUserPageState extends State<AddUserPage> {
   }
 
   Future<void> _pickImage() async {
-    File selectedFile = await FilePicker.getFile(type: FileType.image);
+    FilePickerResult selectedFile = await FilePicker.platform.pickFiles(type: FileType.image);
     if (selectedFile != null && this.mounted) {
       setState(() {
-        profilePic = selectedFile;
+        profilePic =File(selectedFile.files.first.path);
       });
     }
   }
@@ -73,15 +73,14 @@ class _AddUserPageState extends State<AddUserPage> {
         uploadingPop.show();
         String profile_pic_url = null;
         if (profilePic != null) {
-          StorageReference storageReference = FirebaseStorage.instance
-              .ref()
-              .child(
+              FirebaseStorage storage = FirebaseStorage.instance;
+              Reference ref = storage.ref().child(
                   'AvanaFiles/profilepics/' + profilePic.path.split("/").last);
-          StorageUploadTask uploadTask = storageReference.putFile(profilePic);
-          await uploadTask.onComplete;
-          profile_pic_url = await storageReference.getDownloadURL();
-        }
-        await Firestore.instance.collection("userdata").add({
+              UploadTask uploadTask = ref.putFile(profilePic);
+              uploadTask.then((res) {
+                  res.ref.getDownloadURL().then((value) async {
+                    profile_pic_url=value;
+                    await FirebaseFirestore.instance.collection("userdata").add({
           "username": userName.text,
           "email": emailId.text,
           "password": password.text,
@@ -106,6 +105,37 @@ class _AddUserPageState extends State<AddUserPage> {
         }
         uploadingPop.hide();
         Navigator.pop(context);
+                  } );
+              });
+        }else{
+await FirebaseFirestore.instance.collection("userdata").add({
+          "username": userName.text,
+          "email": emailId.text,
+          "password": password.text,
+          "isactive": isActive,
+          "userrole": int.parse(userRole),
+          "membershipdate": new DateTime.now().millisecondsSinceEpoch,
+          "description": description.text,
+          "hospital": hospital.text,
+          "city": city.text,
+          "region": region,
+          "profile_pic_url": profile_pic_url,
+        });
+
+        userName.clear();
+        emailId.clear();
+        password.clear();
+        description.clear();
+        if (this.mounted) {
+          setState(() {
+            loading = false;
+          });
+        }
+        uploadingPop.hide();
+        Navigator.pop(context);
+        }
+
+        
       }
     } catch (Exception) {
       setState(() {
