@@ -221,32 +221,29 @@ class _HomePageState extends State<HomePage> {
             items: <BottomNavigationBarItem>[
               BottomNavigationBarItem(
                 icon: Icon(Icons.home),
-                title: Text('Home'),
+                label: 'Home',
               ),
               BottomNavigationBarItem(
                 icon: Icon(Icons.rss_feed),
-                title: Text('Feed'),
+                label: 'Feed',
               ),
               BottomNavigationBarItem(
                 icon: Icon(Icons.message),
-                title: Text('Message'),
+                label: 'Message',
               ),
               BottomNavigationBarItem(
                 icon: Icon(Icons.image),
-                title: Text('Resources'),
+                label:'Resources',
               ),
               BottomNavigationBarItem(
                 icon: Icon(Icons.supervisor_account),
-                title: Text(
-                  'Users',
-                ),
+                label:  'Users',
+                
               ),
               BottomNavigationBarItem(
                 icon: Icon(Icons.verified_user),
-                title: Text(
-                  'Faculties',
-                ),
-              )
+                label: 'Faculties',
+                              )
             ],
             currentIndex: _selectedIndex,
             backgroundColor: Colors.white,
